requirements.txt
psycopg2-binary>=2.9.3
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=0.21.0

# requirements-dev.txt
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.10.0
isort>=5.10.1
mypy>=0.991
pylint>=2.15.0
httpx>=0.23.0